+++ crates/nostr/src/nips/mod.rs
@@ -50,4 +50,5 @@ pub mod nip90;
 pub mod nip94;
 #[cfg(all(feature = "std", feature = "nip96"))]
 pub mod nip96;
+#[cfg(feature = "nip98")]
 pub mod nip98;