+++ crates/nostr/src/prelude.rs
@@ -71,6 +71,7 @@ pub use crate::nips::nip90::{self, *};
 pub use crate::nips::nip94::{self, *};
 #[cfg(all(feature = "std", feature = "nip96"))]
 pub use crate::nips::nip96::{self, *};
+#[cfg(feature = "nip98")]
 pub use crate::nips::nip98::{self, *};
 pub use crate::signer::{self, *};
 pub use crate::types::*;