+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 
 * nostr: update `ClientMessage::neg_open` signature ([Yuki Kishimoto])
 * nostr: remove `ClientMessage::neg_open_deprecated` ([Yuki Kishimoto])
+* nostr: add `nip98` feature ([Yuki Kishimoto])
 * pool: change `Relay::connect` method signature ([Yuki Kishimoto])
 * ffi: remove embedded tor client ([Yuki Kishimoto])
 * ffi: change `UnsignedEvent::tags` output ([Yuki Kishimoto])
@@ -64,6 +65,7 @@
 * nostr: add `Kind::PeerToPeerOrder` variant ([Francisco Calderón])
 * nostr: add `CowTag` ([Yuki Kishimoto])
 * nostr: add `EventBorrow` ([Yuki Kishimoto])
+* nostr: add `HttpData::to_authorization` ([Yuki Kishimoto])
 * database: add `Events::first_owned` and `Events::last_owned` ([Yuki Kishimoto])
 * database: impl `FlatBufferDecodeBorrowed` for `EventBorrow` ([Yuki Kishimoto])
 * pool: add `Relay::try_connect` ([Yuki Kishimoto])